Project_1_Image_Classification_of_Animals_v2/streamlit_app/requirements.txt
streamlit>=1.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow
numpy
matplotlib
captum